42 Bottom Reinforcements on a 235x235 bed
Daniel Lee (West San Jose)  10:29 PM
For Advanced Users that know their Ender-3 well: This is a 42 bottom reinforcements crammed in on a 235x235 bed.  I print it on my heavily modded Ender-3, with custom Marlin 2.0 firmware to override the stock 220x220 bed limitation, and I’ve tuned my home offsets so slicer knows exact center of bed within one mm.  You literally have a mm or two to spare on the very edges of the bed for this one, and need perfect bed level (or good probing, I have a 7x7 grid)